feat: refactor Hero component import path and add Hero section component
